requirements.txt
Django==4.0.4
pip==22.1.2
django-ckeditor~=6.4.1
Pillow==9.1.1
pip==22.1.2
psycopg2-binary==2.9.3
